+++ spec/models/post_spec.rb
@@ -3,9 +3,21 @@
 RSpec.describe Post do
   include_examples "customizable"
 
-  context "active_fields configuration" do
-    it "allows only provided types" do
-      expect(described_class.active_fields_config.types).to eq(%i[boolean date_array decimal ip ip_array])
+  context "methods" do
+    describe "##allowed_field_type_names" do
+      subject(:call_method) { described_class.allowed_field_type_names }
+
+      it "contains only provided field types" do
+        expect(call_method).to eq(%i[boolean date_array decimal ip ip_array])
+      end
+    end
+
+    describe "##allowed_field_class_names" do
+      subject(:call_method) { described_class.allowed_field_class_names }
+
+      it "contains only provided field class names" do
+        expect(call_method).to eq(ActiveFields.config.fields.values_at(*%i[boolean date_array decimal ip ip_array]))
+      end
     end
   end
 end